Many changes (too many)
The idea is to keep alive the cosmological functions that live
outside the class Cosmology. This forced several modifications
and the option to take a instance of the Cosmology class as a
parameter. In this way they can use the correct growth factors,
growth rate etc... for non flat LCDM cosmologies.
Now most of those functions can do only flat LCDM, while others
that are used for the Linear Point are able to use whatever
cosmology. One possibility would be to remove these functions all
together and only use them inside the Class. In this way
everything will be simpler and consistent. Only issue: need to
always instantiate in order to perform any computation.